% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_freqs}
\alias{allele_freqs}
\title{allele frequency of SNPs in Arabidopsis}
\format{
## `who`
dataframe of 6463 rows and 13 columns
}
\source{
https://plantmethods.biomedcentral.com/articles/10.1186/s13007-014-0041-7
}
\usage{
allele_freqs
}
\description{
dataset of Arabidopsis mutant SNP positions
and SNP effect annotations, including alternate alleles,
transition types and synonymous effects,
}
\keyword{datasets}
