% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arab_infection}
\alias{arab_infection}
\title{arabidopsis mock and infection data matrix}
\format{
## `who`
numeric matrix of 157332 elements in 6 columns and 26222 rows
}
\source{
www.ncbi.nlm.nih.gov/geo/
}
\usage{
arab_infection
}
\description{
Matrix of 6 expression count data from arabidopsis leaves
infected with P.syringae hrcC mutant or a mock infection
}
\keyword{datasets}
