% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arab_infection.rse}
\alias{arab_infection.rse}
\title{arabidopsis mock and infection RangedSummarizedExperiment}
\format{
## `who`
Large RangedSummarizedExperiment object
}
\source{
www.ncbi.nlm.nih.gov/geo/
}
\usage{
arab_infection.rse
}
\description{
RSE of 6 expression count data from arabidopsis leaves
infected with P.syringae hrcC mutant or a mock infection
}
\keyword{datasets}
