% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{at_tf_gex_tidy}
\alias{at_tf_gex_tidy}
\title{arabidopsis transcription factor expression in seedling, root and flowers}
\format{
## `who`
dataframe with 7 columns and 73389 rows
}
\source{
www.ebi.ac.uk/gxa/experiments/E-GEOD-53197/
}
\usage{
at_tf_gex_tidy
}
\description{
tidy format data frame of transcription factor expression levels in various arabidopsis
tissues. Annotated with gene lengths and transcription factor families
}
\keyword{datasets}
