% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath_seq_names}
\alias{ath_seq_names}
\title{unformatted gene headers}
\format{
## `who`
character vector of info for ten genes
}
\source{
www.arabidipsis.org
}
\usage{
ath_seq_names
}
\description{
unformatted gene headers
}
\keyword{datasets}
