% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath_ts}
\alias{ath_ts}
\title{time series gene expression data for 100 genes}
\format{
## `who`
dataframe of 400 rows and 8 variables
}
\source{
figshare.com/articles/dataset/NetSeekR_data/17447384
}
\usage{
ath_ts
}
\description{
genes from Arabidopsis challenged
with flg22 over 12h
}
\keyword{datasets}
