% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_df}
\alias{census_df}
\title{data frame of census data for Chapter 2 recipe 2}
\format{
## `who`
dataframe with 365 rows and 24 columns
}
\source{
https://www.ons.gov.uk/
}
\usage{
census_df
}
\description{
readr parsed data from UK National Census 2021
}
\keyword{datasets}
