% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chapter_6_recipe_3}
\alias{chapter_6_recipe_3}
\title{Download data for Ch 6 Recipe 3}
\usage{
chapter_6_recipe_3()
}
\description{
download all fastq files and transcript fasta files from
https://doi.org/10.5281/zenodo.1294051
}
