% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compost}
\alias{compost}
\title{model compost and yield data}
\format{
## `who`
dataframe of 32 rows and 3 columns
}
\usage{
compost
}
\description{
dataframe of size by growth supplement and
compost type in plant experiments
}
\keyword{datasets}
