% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_dataframe}
\alias{count_dataframe}
\title{modencodefly transcript count data for Chapter 6 recipe 1}
\format{
## `who`
A data frame with 14869 rows and 148 columns
}
\source{
www.modencode.org
}
\usage{
count_dataframe
}
\description{
A count dataframe of the modencodefly data from the NHGRI encyclopedia
of DNA elements for Drosophila
}
\keyword{datasets}
