% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_scores}
\alias{disease_scores}
\title{model plant disease scoring data}
\format{
## `who`
dataframe of 9 rows and 3 columns
}
\usage{
disease_scores
}
\description{
dataframe of disease score by strain and replicate
}
\keyword{datasets}
