% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression}
\alias{expression}
\title{model expression data}
\format{
## `who`
dataframe of 54 rows and 4 columns
}
\usage{
expression
}
\description{
dataframe of computer generated expression values
for 6 genes, in three treatments with three replicates
}
\keyword{datasets}
