% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fission_ge}
\alias{fission_ge}
\title{gene expression data from fission yeast wt vs mut}
\format{
## `who`
dataframe of 7039 rows and 7 columns
}
\source{
Leong et al 2014, Nat Commun. PMID: 24853205.
}
\usage{
fission_ge
}
\description{
gene expression data from fission yeast wt vs mut
}
\keyword{datasets}
