% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_from_PR}
\alias{function_from_PR}
\title{a function that exists only in the 'example_devtools_branch'}
\usage{
function_from_PR()
}
\description{
a function that exists only in the 'example_devtools_branch'
}
