% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_gene_exp}
\alias{mo_gene_exp}
\title{Gene expression values for ten M oryzae genes}
\format{
## `who`
dataframe with 10 rows and 1 column
}
\usage{
mo_gene_exp
}
\description{
Gene expression values for ten M oryzae genes
}
\keyword{datasets}
