% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_go_acc}
\alias{mo_go_acc}
\title{GO accession numbers for M oryzae genes}
\format{
## `who`
dataframe of 59619 rows and 2 columns
}
\source{
http://fungi.ensembl.org
}
\usage{
mo_go_acc
}
\description{
GO accession numbers for M oryzae genes
}
\keyword{datasets}
