% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_go_evidence}
\alias{mo_go_evidence}
\title{GO evidence codes for M oryzae genes}
\format{
## `who`
dataframe of 59619 rows and 4 columns
}
\source{
http://fungi.ensembl.org
}
\usage{
mo_go_evidence
}
\description{
GO evidence codes for M oryzae genes
}
\keyword{datasets}
