% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_terms}
\alias{mo_terms}
\title{GO terms for M oryzae genes}
\format{
## `who`
dataframe of 59619 rows and 5 columns
}
\source{
http://fungi.ensembl.org
}
\usage{
mo_terms
}
\description{
GO terms for M oryzae genes
}
\keyword{datasets}
