% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modencodefly.eset}
\alias{modencodefly.eset}
\title{modencode transcript data as an eset object for Chapter 6 recipe 1}
\format{
## `who`
A Large ExpressionSet object
}
\source{
www.modencode.org
}
\usage{
modencodefly.eset
}
\description{
A count object typical in Bioconductor analyses of
the modencodefly data from the NHGRI encyclopedia
of DNA elements for Drosophila
}
\keyword{datasets}
