% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{package_sgfc}
\alias{package_sgfc}
\title{precomputed sgfc for Chapter 6 Recipe 5}
\format{
## `who`
SGFeatureCounts object
}
\usage{
package_sgfc
}
\description{
An SGFeatureCounts object that results from Chapter 6 Recipe 5 step 4
but can take a very long time to compute.
}
\keyword{datasets}
