% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_data}
\alias{pheno_data}
\title{modencode phenotype data for Chapter 6 recipe 1}
\format{
## `who`
A data frame with 147 rows and 1 column
}
\source{
www.modencode.org
}
\usage{
pheno_data
}
\description{
modencode phenotype data for Chapter 6 recipe 1
}
\keyword{datasets}
