% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_growth_two}
\alias{plant_growth_two}
\title{plant growth data on two groups}
\format{
## `who`
dataframe of 20 rows and 10 columns
}
\source{
"R"
}
\usage{
plant_growth_two
}
\description{
dataset of weight and treatement measurements
derived from built-in PlantGrowth data with
just control and trt1 data
}
\keyword{datasets}
