% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{side_effects}
\alias{side_effects}
\title{Example tibble of side effect frequencies}
\format{
A tibble with 9 rows and 3 columns:
  \describe{
    \item{\code{side_effect}}{character: the type of side effect experienced (none, headache, or nausea)}
    \item{\code{treatment}}{character: the treatment administered (regime_a, regime_b, or regime_c)}
    \item{\code{frequency}}{numeric: the frequency of the side effect for the given treatment}
  }
}
\usage{
side_effects
}
\description{
This tibble contains the frequency of side effects for three different treatments (regime_a, regime_b, and regime_c). Each row represents a different observation with three variables: the type of side effect experienced (\code{side_effect}), the treatment administered (\code{treatment}), and the frequency of the side effect (\code{frequency}).
}
\keyword{datasets}
