% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{data frame of treatment data for Chapter 2 recipe 3}
\format{
## `who`
dataframe with 18 rows and 4 columns
}
\usage{
treatments
}
\description{
made up treatment data to show splitting condensed columns
}
\keyword{datasets}
